requirements.txt
asgiref==3.4.1
Django==3.2.5
djangorestframework==3.12.4
gunicorn==20.1.0
python-dotenv==0.18.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
